Graphics Revamp of Spell Cards
Enhanced the detail of the cards by adding grey shading and a faint shine to the symbols. I also made the font of the card names easier to read by thickening the letters and adding an extra, lighter shade of red. Some cards have also had additional strokes to accentuate the symbols.